pyproject.toml
[project]
name = "auth-mcp-remote"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "certifi>=24.0.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.0",
    "httpx>=0.28.1",
    "mcp[cli]>=1.18.0",
    "pydantic>=2.12.3",
    "python-dotenv>=1.1.1",
    "pyjwt[crypto]>=2.8.0",
    "requests>=2.32.0",
    "youtube-transcript-api>=1.2.3",
]
